feat(web): update routing types and add safe-area utility
- Changed import path for routing types in next-env.d.ts to use the development directory.
- Added a new CSS utility class for safe-area insets to globals.css, enhancing layout support for notched devices.
- Introduced a new MarketingHome component for improved marketing page structure and navigation.
- Updated app-shell and app-ui components for better layout management and transition effects.
- Added formatting utilities for event data in format.ts to streamline event display logic.
