+++ apps/web/app/globals.css
@@ -93,6 +93,11 @@ img {
   max-width: 100%;
 }
 
+/* Safe-area utility for notched devices (iPhone home indicator) */
+.safe-area-inset-bottom {
+  padding-bottom: env(safe-area-inset-bottom, 0);
+}
+
 @keyframes imageCycle {
   0%,
   29% {
