+++ apps/web/lib/format.ts
@@ -1,0 +1,35 @@
+import type { EventCard, AdminMarketSummary } from '@aventi/contracts';
+
+export function formatDateTime(value?: string | null) {
+  if (!value) return 'Not recorded';
+  return new Intl.DateTimeFormat(undefined, {
+    month: 'short',
+    day: 'numeric',
+    hour: 'numeric',
+    minute: '2-digit',
+  }).format(new Date(value));
+}
+
+export function formatEventTime(startsAt: string) {
+  return new Intl.DateTimeFormat(undefined, {
+    weekday: 'short',
+    month: 'short',
+    day: 'numeric',
+    hour: 'numeric',
+    minute: '2-digit',
+  }).format(new Date(startsAt));
+}
+
+export function formatPrice(event: EventCard) {
+  if (event.isFree) return 'Free';
+  return event.priceLabel || event.ticketOffers?.[0]?.priceLabel || 'Check venue';
+}
+
+export function marketStatus(market: AdminMarketSummary) {
+  if (market.lastError) return 'attention';
+  if (market.scanLockUntil && new Date(market.scanLockUntil).getTime() > Date.now()) return 'warming';
+  if (market.lastTargetedRequestedAt && market.lastTargetedRequestedAt !== market.lastTargetedCompletedAt) {
+    return 'targeted_warming';
+  }
+  return 'ready';
+}
